refactor(helpers): move transition factories to components
fixes #9824
